Improved tolerance for high API latanzy
However, this is simply not going to work with API latencies exceeding 60 seconds.

But this was enough to verify that the trip endpoint, as of now (2024), actually saves the correct delays, rather than in 2021, where it often removed delays.

This means I'll now rewrite the processor to use the trip API because then API latency will not cause it to fail or make data inaccurate. Data might just be delayed.